# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  codex.sourceCandidate:
    properties:
      connector:
        type: string
      id:
        type: string
      kind:
        type: string
      label:
        type: string
      uri:
        type: string
    type: object
  codex.sourceDiscoveryResponse:
    properties:
      connector:
        type: string
      provider:
        type: string
      sources:
        items:
          $ref: '#/definitions/codex.sourceCandidate'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Local-first pipeline API for ingesting and reasoning over engineering
    context.
  title: ContextOS API
  version: "1.0"
paths:
...
  /codex/login:
    post:
      description: Runs `codex login --device-auth` and streams log lines as SSE events.
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Trigger Codex device login
      tags:
      - codex
  /codex/plugin-reauth:
    post:
      description: Removes then re-adds the named plugin to trigger a fresh OAuth
        flow.
      parameters:
      - description: 'Plugin short name: github, atlassian-rovo, jira, or slack'
        in: query
        name: plugin
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Re-authenticate a Codex plugin
      tags:
      - codex
  /codex/sources:
    get:
      description: Uses Codex plugins to list readable sources for github, jira, slack,
        notion, googledrive, or sharepoint.
      parameters:
      - description: 'Connector: github, jira, slack, notion, googledrive, or sharepoint'
        in: query
        name: connector
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codex.sourceDiscoveryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Codex-accessible sources
      tags:
      - codex
  /codex/status:
    get:
      description: Returns Codex CLI version, login status, and installed plugin list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Codex CLI status
      tags:
      - codex
...
swagger: "2.0"